Enhancement: Re-estructuración del codigo para dao implementación de interfaces Se implemento las interfaces de negocio para los daos
