+++ functions/src/dao/impl/addressDaoImpl.ts
@@ -1,0 +1,74 @@
+import * as log4js from 'log4js';
+import Mysql from "../../utils/mysql";
+import { AbstractDao } from "../abstractDao";
+import { AddressDao } from '../addressDao';
+
+const logger = log4js.getLogger('sumagro.dao.impl.AddressDaoImpl');
+
+export class AddressDaoImpl extends AbstractDao implements AddressDao {
+
+    constructor(db: Mysql) {
+        super(db);
+    }
+
+    async getAddress(address: any){
+        logger.info(`Dao: statting getAdress`);
+        logger.info(`Dao: address: ${address}`);
+        let sql=`SELECT id FROM address where localidad='${address}'`;
+        logger.info(`Dao: ending getAdress`);
+        return await this.db.query(sql);
+    }
+
+    async getLocalidadAddressById(addressId: number){
+        logger.info('DAO:Method getLocalidadAddressById Startting');
+        let sql= `SELECT localidad FROM address WHERE id=${addressId}`;
+        logger.debug('DAO: Method getLocalidadAddressById Ending');
+        return await this.db.query(sql);
+    }
+
+    async CreateAddress(address: any){
+        logger.info('DAO: Method CreateAddress Starting');
+        let sql = `INSERT INTO address(calle,numero,ciudad,localidad,municipio) VALUES('${address.street}','${address.number}',"${address.city}","${address.location}","${address.municipality}")`;
+        logger.debug('DAO: Method CreateAddress Ending');
+        return await this.db.query(sql);
+    }
+
+    async createAddressByLocalidad(localidad: any){
+        logger.info('DAO: Method createAddressByLocalidad Startting');
+        let sql= `INSERT INTO address(localidad) VALUES('${localidad}')`;
+        let res:any= await this.db.query(sql);
+        logger.debug('DAO: Method createAddressByLocalidad Ending');
+        return res.insertId;
+    }
+
+    async getAddresByAttributes(address: any){
+        logger.info('DAO: Method CreateAddress Starting');
+        let sql = `SELECT * FROM address WHERE calle = '${address.street}' AND numero = '${address.number}' AND ciudad = "${address.city}" AND localidad = "${address.location}" AND municipio = "${address.municipality}"`;
+       logger.info(`DAO: sql ${sql}`);
+        logger.debug('DAO: Method CreateAddress Ending');
+        return await this.db.query(sql);
+    }
+    async getAddressById(addressId: number){
+        logger.info('DAO: Method getAddressById Starting');
+        let sql= `SELECT * FROM address WHERE id= ${addressId}`;
+        logger.debug('DAO: Method getAddressById Ending');
+    return await this.db.query(sql);
+    }
+
+    async deleteAddresById(addressId: string){
+        logger.info('DAO: Method deleteAddresById Starting');
+        let sql = `DELETE FROM address WHERE id = ${addressId}`;
+        logger.debug('DAO: Method deleteAddresById Ending');
+        return await this.db.query(sql);
+        
+    }
+
+    async getIdAddresByAttributes(address: any){
+        logger.info('DAO: Method CreateAddress Starting');
+        let sql = `SELECT * FROM address WHERE calle = '${address.calle}' AND numero = '${address.numero}' AND ciudad = "${address.ciudad}" AND localidad = "${address.localidad}" AND municipio = "${address.municipio}"`;
+       logger.info(`DAO: sql ${sql}`);
+        logger.debug('DAO: Method CreateAddress Ending');
+        return await this.db.query(sql);
+    }
+
+}
